+++ src/main/java/reactor/core/publisher/LambdaFirstSubscriber.java
@@ -166,7 +166,7 @@ final class LambdaFirstSubscriber<T>
 
 	@Override
 	public long getCapacity() {
-		return Long.MAX_VALUE;
+		return Integer.MAX_VALUE;
 	}
 
 	@Override
